Change place of tracking chosen cells
Before the change chosen cells were stored inside an object in gameFlow
module. I decided that a better place to do it will be in players
objects, so now objects have array to store their choices. After this
change it was necessary to check which player plays as x, so new
function was added to gameFlow module that returns which object plays as
x and which as o.